Rename kprovex -> kprove
