+++ kernel/src/main/java/org/kframework/kprove/KProveOptions.java
@@ -1,12 +1,11 @@
 // Copyright (c) K Team. All Rights Reserved.
 
-package org.kframework.kprovex;
+package org.kframework.kprove;
 
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.ParametersDelegate;
 import org.kframework.unparser.PrintOptions;
 import org.kframework.main.GlobalOptions;
-import org.kframework.utils.errorsystem.KEMException;
 import org.kframework.utils.file.FileUtil;
 import org.kframework.utils.inject.RequestScoped;
 import org.kframework.utils.options.BackendOptions;
